+++ src/grasplan/tools/common.py
